requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic[email]
pdfplumber
python-docx
openai-agents
json-repair
loguru